src/example-1-instructor-basics/pyproject.toml
[project]
name = "example-1-instructor-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fireworks-ai>=0.15.12",
    "instructor[fireworks]>=1.7.2",
    "orjson>=3.10.15",
]
